services/webapi-service/pyproject.toml
[project]
name = "bombermanonline"
version = "0.1.0"
description = "Bomberman Online - многопользовательская игра в Bomberman"
requires-python = ">=3.12"
dependencies = [
    "python-engineio>=4.12.0",
    "python-socketio>=5.13.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "aiohttp>=3.9.5",
    "pydantic-settings>=2.9.1",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.40",
    "numpy>=2.2.5",
    "nats-py>=2.10.0",
]